fix issue #2.  grunt-file-creator prints an error when using verbose mode
grunt has a special parsing mechanism for displaying 'file' options in verbose mode.  Unfortunately, this function cannot handle file params that are functions.

The clearest fix is to simply specify the options directly (rather than use a 'files' property) so that the special parsing method is not used.  There's no real reason to use 'files' as the property anyway and, besides, the spec is shorter.

OLD:
      "basic": {
        files: {
          "tmp/basic.txt": function(fs, fd, done) {
            fs.writeSync(fd, 'some basic text');
            done();
          }
        }
      },

NEW:
      "basic": {
        "tmp/basic.txt": function(fs, fd, done) {
          fs.writeSync(fd, 'some basic text');
          done();
        }
      },